removeAssignedUserFromTemplate and new shared notifications file
